+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the "vscode-java-test" extension will be documented in th
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+* Fix JUnit test discovery for projects using preview features from an older Java release.
+
 ## 0.46.0
 ## What's Changed
 * feat - Run multiple selected test methods in a single JVM in https://github.com/microsoft/vscode-java-test/pull/1862